\name{ExcelScript-package}
\alias{ExcelScript-package}
\alias{ExcelScript}
\docType{package}
\title{
\packageTitle{ExcelScript}
}
\description{
\packageDescription{ExcelScript}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ExcelScript}
\packageIndices{ExcelScript}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ExcelScript}

Maintainer: \packageMaintainer{ExcelScript}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
